LayoutTests/fast/table/auto-layout-width-percentage-exceed-hundred-expected.txt

PASS document.getElementById("onlyTable").offsetWidth is 1000000
PASS successfullyParsed is true

TEST COMPLETE
